refactor: extract widget-placement state machine from OrgDashboardPage
index.tsx packed a 13-variable state machine driving drag, keyboard, and
click/touch placement of dashboard widgets into a single 1,033-line
component. Split it into useWidgetPlacement (the placement state machine
and its handlers) and EditableWidgetTile (the presentational tile), leaving
index.tsx to orchestrate layout/save state and wire the two together.

Fixes #867
